requirements.txt
dash==2.7.0
dash-bootstrap-components==1.2.1
loguru==0.6.0
plotly==5.11.0
requests==2.28.1
numpy==1.23.4
python-dateutil==2.8.2
dash-extensions==0.1.7
Flask-Caching==2.0.1
pandas==1.5.1
PyMySQL==1.0.2
statsmodels==0.13.2
nbformat==5.7.0
plotly-calplot==0.1.13